Minor ChatView improvements
- Minor improvements to ChatBubble
- Minor improvements to ChatView
