+++ packages/ui/src/hooks/useSideMenu.ts
@@ -11,8 +11,9 @@ export const useSideMenu = (
     const [isOpen, setIsOpen] = useState<boolean>(isMobile ? false : defaultOpenState);
     useEffect(() => {
         const sideMenuSub = PubSub.get().subscribeSideMenu((data) => {
-            console.log("got side menu event", data);
+            console.log("got side menu event", data, id);
             if (data.id !== id) return;
+            console.log('setting side menu state', data.id, data.isOpen)
             setIsOpen(data.isOpen);
             setSideMenuState(id, data.isOpen);
         });
